feat: prepare main-cart and mini-cart to introduce in storage
